split InAppMessageView for Base to be reusable by OM
