+++ OptimoveSDK/optimove-sdk/src/main/java/com/optimove/android/optimobile/AnalyticsContract.java
@@ -45,6 +45,7 @@ final class AnalyticsContract {
     static final String EVENT_TYPE_MESSAGE_DELIVERED = "k.message.delivered";
     static final String EVENT_TYPE_MESSAGE_READ = "k.message.read";
     static final String MESSAGE_DELETED_FROM_INBOX = "k.message.inbox.deleted";
+    static final String EVENT_TYPE_OM_INTERCEPTED = "optimove.om.intercepted";
     static final String EVENT_TYPE_DEEP_LINK_MATCHED = "k.deepLink.matched";
     static final String EVENT_TYPE_LOCATION_UPDATED = "k.engage.locationUpdated";
     static final String EVENT_TYPE_ENTERED_BEACON_PROXIMITY = "k.engage.beaconEnteredProximity";
