+++ OptimoveSDK/optimove-sdk/src/main/java/com/optimove/android/optimobile/InAppMessageView.java
@@ -1,46 +1,13 @@
 package com.optimove.android.optimobile;
 
-import android.annotation.SuppressLint;
-import android.annotation.TargetApi;
 import android.app.Activity;
-import android.app.Dialog;
-import android.content.Context;
 import android.content.Intent;
-import android.graphics.Rect;
 import android.net.Uri;
-import android.net.http.SslError;
-import android.os.Build;
-import android.util.DisplayMetrics;
 import android.util.Log;
-import android.util.Pair;
-import android.view.Display;
-import android.view.DisplayCutout;
-import android.view.KeyEvent;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.Window;
-import android.view.WindowInsets;
-import android.view.WindowManager;
-import android.webkit.JavascriptInterface;
-import android.webkit.RenderProcessGoneDetail;
-import android.webkit.SslErrorHandler;
-import android.webkit.WebResourceError;
-import android.webkit.WebResourceRequest;
-import android.webkit.WebResourceResponse;
-import android.webkit.WebSettings;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-import android.widget.ProgressBar;
-import android.widget.RelativeLayout;
 
-import androidx.annotation.AnyThread;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.RequiresApi;
 import androidx.annotation.UiThread;
-
-import com.optimove.android.BuildConfig;
-import com.optimove.android.R;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -49,14 +16,7 @@ import org.json.JSONObject;
 import java.util.ArrayList;
 import java.util.List;
 
-class InAppMessageView extends WebViewClient {
-
-    private enum State {
-        INITIAL,
-        LOADING,
-        READY,
-        DISPOSED
-    }
+class InAppMessageView extends BaseMessageView {
 
     private static final String TAG = InAppMessageView.class.getName();
 
@@ -67,29 +27,6 @@ class InAppMessageView extends WebViewClient {
     private static final String BUTTON_ACTION_REQUEST_APP_STORE_RATING = "requestAppStoreRating";
     private static final String BUTTON_ACTION_PUSH_REGISTER = "promptPushPermission";
 
-    private static final String HOST_MESSAGE_TYPE_PRESENT_MESSAGE = "PRESENT_MESSAGE";
-    private static final String HOST_MESSAGE_TYPE_CLOSE_MESSAGE = "CLOSE_MESSAGE";
-    private static final String HOST_MESSAGE_TYPE_SET_NOTCH_INSETS = "SET_NOTCH_INSETS";
-
-    private static final String JS_NAME = "Android";
-
-    private State state;
-    private boolean pageFinished;
-
-    @NonNull
-    private final Activity currentActivity;
-
-    @Nullable
-    private WebView wv;
-    @Nullable
-    private Dialog dialog;
-    @Nullable
-    private ProgressBar spinner;
-
-    private int prevStatusBarColor;
-    private boolean prevFlagTranslucentStatus;
-    private boolean prevFlagDrawsSystemBarBackgrounds;
-
     @NonNull
     private final InAppMessagePresenter presenter;
     @NonNull
@@ -98,11 +35,14 @@ class InAppMessageView extends WebViewClient {
     private final String region;
 
     @UiThread
-    InAppMessageView(@NonNull InAppMessagePresenter presenter, @NonNull InAppMessage message, @NonNull Activity currentActivity, @NonNull String iarUrl, @Nullable String region) {
-        this.state = State.INITIAL;
-        pageFinished = false;
+    InAppMessageView(@NonNull InAppMessagePresenter presenter,
+                     @NonNull InAppMessage message,
+                     @NonNull Activity currentActivity,
+                     @NonNull String iarUrl,
+                     @Nullable String region) {
+        super(currentActivity);
+
         this.presenter = presenter;
-        this.currentActivity = currentActivity;
         this.currentMessage = message;
         this.region = region;
 
@@ -116,396 +56,6 @@ class InAppMessageView extends WebViewClient {
         }
         currentMessage = message;
         sendCurrentMessageToClient();
-    }
-
-    @UiThread
-    void dispose() {
-        closeDialog(currentActivity);
-        state = State.DISPOSED;
-    }
-
-    @UiThread
-    private void sendCurrentMessageToClient() {
-        if (state == State.READY && pageFinished) {
-            JSONObject content = currentMessage.getContent();
-            if (region != null) {
-                try {
-                    content.put("region", region);
-                } catch (JSONException e) {
-                    Log.w(TAG, "Could not pass region to In-App renderer");
-                }
-            }
-
-            sendToClient(HOST_MESSAGE_TYPE_PRESENT_MESSAGE, content);
-        }
-    }
-
-    @UiThread
-    private void setSpinnerVisibility(int visibility) {
-        if (spinner != null) {
-            spinner.setVisibility(visibility);
-        }
-    }
-
-    @UiThread
-    private void sendToClient(String type, JSONObject data) {
-        if (wv == null) {
-            return;
-        }
-
-        JSONObject j = new JSONObject();
-        try {
-            j.put("data", data);
-            j.put("type", type);
-        } catch (JSONException e) {
-            Log.d(TAG, "Could not create client message");
-            return;
-        }
-
-        String script = "window.postHostMessage(" + j.toString() + ")";
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-            wv.evaluateJavascript(script, null);
-        } else {
-            wv.loadUrl("javascript:" + script);
-        }
-    }
-
-    @UiThread
-    @SuppressWarnings("deprecation")
-    private void setStatusBarColorForDialog(Activity currentActivity) {
-        if (currentActivity == null) {
-            return;
-        }
-
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-            return;
-        }
-
-        Window window = currentActivity.getWindow();
-
-        prevStatusBarColor = window.getStatusBarColor();
-
-        int flags = window.getAttributes().flags;
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.R) {
-            prevFlagTranslucentStatus = (flags & WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS) != 0;
-            window.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
-        }
-
-        prevFlagDrawsSystemBarBackgrounds = (flags & WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS) != 0;
-        window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-        int statusBarColor;
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            statusBarColor = currentActivity.getResources().getColor(R.color.statusBarColorForNotch, null);
-        } else {
-            statusBarColor = currentActivity.getResources().getColor(R.color.statusBarColorForNotch);
-        }
-
-        window.setStatusBarColor(statusBarColor);
-    }
-
-    @UiThread
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-    private void unsetStatusBarColorForDialog(Activity dialogActivity) {
-        if (dialogActivity == null) {
-            return;
-        }
-
-        Window window = dialogActivity.getWindow();
-        window.setStatusBarColor(prevStatusBarColor);
-
-        if (prevFlagTranslucentStatus) {
-            window.addFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
-        }
-
-        if (!prevFlagDrawsSystemBarBackgrounds) {
-            window.clearFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-        }
-    }
-
-    @UiThread
-    private void closeDialog(Activity dialogActivity) {
-        if (dialog != null) {
-            dialog.setOnKeyListener(null);
-            dialog.dismiss();
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                unsetStatusBarColorForDialog(dialogActivity);
-            }
-        }
-
-        if (null != wv) {
-            wv.destroy();
-        }
-
-        dialog = null;
-        wv = null;
-        spinner = null;
-    }
-
-    @SuppressLint({"SetJavaScriptEnabled", "AddJavascriptInterface", "InflateParams"})
-    @UiThread
-    private void showWebView(@NonNull final Activity currentActivity, @NonNull String iarUrl) {
-        try {
-            if (BuildConfig.DEBUG && Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-                WebView.setWebContentsDebuggingEnabled(true);
-            }
-
-            RelativeLayout.LayoutParams paramsWebView = new RelativeLayout.LayoutParams(RelativeLayout.LayoutParams.MATCH_PARENT, RelativeLayout.LayoutParams.MATCH_PARENT);
-            dialog = new Dialog(currentActivity, android.R.style.Theme_Translucent_NoTitleBar_Fullscreen);
-
-            Window window = dialog.getWindow();
-            if (window != null && Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-                WindowManager.LayoutParams windowAttributes = dialog.getWindow().getAttributes();
-                windowAttributes.layoutInDisplayCutoutMode = WindowManager.LayoutParams.LAYOUT_IN_DISPLAY_CUTOUT_MODE_SHORT_EDGES;
-
-                View view = window.getDecorView();
-                view.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_STABLE | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN);
-            }
-
-            LayoutInflater inflater = (LayoutInflater) currentActivity.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-            dialog.setContentView(inflater.inflate(R.layout.optimobile_dialog_view, null), paramsWebView);
-            dialog.setOnKeyListener((dialog, keyCode, event) -> {
-                if (keyCode == KeyEvent.KEYCODE_BACK && event.getAction() != KeyEvent.ACTION_DOWN) {
-                    closeCurrentMessage();
-                }
-                return true;
-            });
-
-            wv = dialog.findViewById(R.id.optimobile_webview);
-            spinner = dialog.findViewById(R.id.optimobile_progressBar);
-
-            if (null == wv || null == spinner) {
-                dispose();
-                return;
-            }
-
-            int cacheMode = WebSettings.LOAD_DEFAULT;
-            if (BuildConfig.DEBUG) {
-                cacheMode = WebSettings.LOAD_NO_CACHE;
-            }
-            wv.getSettings().setCacheMode(cacheMode);
-
-            wv.setBackgroundColor(android.graphics.Color.TRANSPARENT);
-
-            WebSettings settings = wv.getSettings();
-            settings.setJavaScriptEnabled(true);
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
-                settings.setMediaPlaybackRequiresUserGesture(false);
-            }
-
-            wv.addJavascriptInterface(this, JS_NAME);
-            wv.setWebViewClient(this);
-
-            dialog.show();
-            setSpinnerVisibility(View.VISIBLE);
-            wv.loadUrl(iarUrl);
-            state = State.LOADING;
-        } catch (Exception e) {
-            Optimobile.log(TAG, e.getMessage());
-        }
-    }
-
-    private void closeCurrentMessage() {
-        sendToClient(HOST_MESSAGE_TYPE_CLOSE_MESSAGE, null);
-        InAppMessageService.handleMessageClosed(currentActivity, currentMessage);
-    }
-
-    @Override
-    public void onPageFinished(WebView view, String url) {
-        view.setBackgroundColor(android.graphics.Color.TRANSPARENT);
-        setStatusBarColorForDialog(currentActivity);
-        pageFinished = true;
-
-        sendCurrentMessageToClient();
-
-        super.onPageFinished(view, url);
-    }
-
-    @Override
-    @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-    public void onReceivedHttpError(WebView view, WebResourceRequest request, WebResourceResponse errorResponse) {
-        super.onReceivedHttpError(view, request, errorResponse);
-
-        String url = request.getUrl().toString();
-
-        try {
-            // Only consider handling for failures of our renderer assets
-            // 3rd-party fonts/images etc. shouldn't trigger this
-            String iarBaseUrl = Optimobile.urlForService(UrlBuilder.Service.IAR, "");
-            if (!url.startsWith(iarBaseUrl)) {
-                return;
-            }
-
-            // Cached index page may refer to stale JS/CSS file hashes
-            // Evict the cache to allow next presentation to re-fetch
-            if (404 == errorResponse.getStatusCode()) {
-                view.clearCache(true);
-            }
-
-            closeDialog(currentActivity);
-        } catch (Optimobile.PartialInitialisationException e) {
-            Optimobile.log(TAG, "Cannot handle HTTP error: credentials not yet available");
-        }
-    }
-
-    @Override
-    public void onReceivedSslError(WebView view, SslErrorHandler handler, SslError error) {
-        handler.cancel();
-        closeDialog(currentActivity);
-    }
-
-    @Override
-    public boolean onRenderProcessGone(WebView view, RenderProcessGoneDetail detail) {
-        closeDialog(currentActivity);
-
-        // Allow app to keep running, don't terminate
-        return true;
-    }
-
-    @Override
-    public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
-        Optimobile.log(TAG, "Error code: " + errorCode + ". " + description + " " + failingUrl);
-
-        String extension = failingUrl.substring(failingUrl.length() - 4);
-        boolean isVideo = extension.matches(".mp4|.m4a|.m4p|.m4b|.m4r|.m4v");
-        if (errorCode == -1 && "net::ERR_FAILED".equals(description) && isVideo) {
-            // This is a workaround for a bug in the WebView.
-            // See these chromium issues for more context:
-            // https://bugs.chromium.org/p/chromium/issues/detail?id=1023678
-            // https://bugs.chromium.org/p/chromium/issues/detail?id=1050635
-
-            //We encountered the issue only with some (and not other) videos, but possibly not limited to other file types
-            return;
-        }
-
-        closeDialog(currentActivity);
-    }
-
-    @Override
-    @TargetApi(Build.VERSION_CODES.M)
-    public void onReceivedError(WebView view, WebResourceRequest req, WebResourceError rerr) {
-        onReceivedError(view, rerr.getErrorCode(), rerr.getDescription().toString(), req.getUrl().toString());
-    }
-
-    @JavascriptInterface
-    @AnyThread
-    public void postClientMessage(String msg) {
-        String messageType;
-        JSONObject data;
-        try {
-            JSONObject message = new JSONObject(msg);
-            messageType = message.getString("type");
-            data = message.optJSONObject("data");
-        } catch (JSONException e) {
-            Log.d(TAG, "Incorrect message format: " + msg);
-            return;
-        }
-
-        switch (messageType) {
-            case "READY":
-                currentActivity.runOnUiThread(() -> {
-                    if (state == State.DISPOSED) {
-                        return;
-                    }
-
-                    state = State.READY;
-
-                    maybeSetNotchInsets(currentActivity);
-                    sendCurrentMessageToClient();
-                });
-                return;
-            case "MESSAGE_OPENED":
-                currentActivity.runOnUiThread(() -> {
-                    if (state == State.DISPOSED) {
-                        return;
-                    }
-
-                    setSpinnerVisibility(View.GONE);
-                    InAppMessageService.handleMessageOpened(currentActivity, currentMessage);
-                });
-                return;
-            case "MESSAGE_CLOSED":
-                currentActivity.runOnUiThread(presenter::messageClosed);
-                return;
-            case "EXECUTE_ACTIONS":
-                if (null == data) {
-                    return;
-                }
-                List<ExecutableAction> actions = this.parseButtonActionData(data);
-                currentActivity.runOnUiThread(() -> this.executeActions(currentActivity, actions));
-                return;
-            default:
-                Log.d(TAG, "Unknown message type: " + messageType);
-        }
-    }
-
-    @UiThread
-    private void maybeSetNotchInsets(Context context) {
-        if (dialog == null) {
-            return;
-        }
-
-        Window window = dialog.getWindow();
-        if (window == null || Build.VERSION.SDK_INT < Build.VERSION_CODES.P) {
-            return;
-        }
-
-        WindowInsets insets = window.getDecorView().getRootWindowInsets();
-        if (insets == null) {
-            return;
-        }
-
-        DisplayCutout displayCutout = insets.getDisplayCutout();
-        if (displayCutout == null) {
-            return;
-        }
-
-        List<Rect> cutoutBoundingRectangles = displayCutout.getBoundingRects();
-        if (cutoutBoundingRectangles.size() == 0) {
-            return;
-        }
-
-        Pair<Boolean, Boolean> notchPositions = determineNotchPositions(window, cutoutBoundingRectangles);
-        float density = context.getResources().getDisplayMetrics().density;
-
-        JSONObject notchData = new JSONObject();
-        try {
-            notchData.put("hasNotchOnTheLeft", notchPositions.first);
-            notchData.put("hasNotchOnTheRight", notchPositions.second);
-            notchData.put("insetTop", displayCutout.getSafeInsetTop() / density);
-            notchData.put("insetRight", displayCutout.getSafeInsetRight() / density);
-            notchData.put("insetBottom", displayCutout.getSafeInsetBottom() / density);
-            notchData.put("insetLeft", displayCutout.getSafeInsetLeft() / density);
-
-            sendToClient(HOST_MESSAGE_TYPE_SET_NOTCH_INSETS, notchData);
-        } catch (JSONException e) {
-            Optimobile.log(TAG, e.getMessage());
-        }
-    }
-
-    @UiThread
-    private Pair<Boolean, Boolean> determineNotchPositions(Window window, List<Rect> cutoutBoundingRectangles) {
-        Display display = window.getWindowManager().getDefaultDisplay();
-        DisplayMetrics outMetrics = new DisplayMetrics();
-        display.getMetrics(outMetrics);
-
-        boolean hasNotchOnTheRight = false;
-        boolean hasNotchOnTheLeft = false;
-        for (Rect rect : cutoutBoundingRectangles) {
-            if (rect.top == 0) {
-                if (rect.left > outMetrics.widthPixels - rect.right) {
-                    hasNotchOnTheRight = true;
-                } else if (rect.left < outMetrics.widthPixels - rect.right) {
-                    hasNotchOnTheLeft = true;
-                }
-            } else if (rect.right >= outMetrics.widthPixels) {
-                hasNotchOnTheRight = true;
-            } else if (rect.left == 0) {
-                hasNotchOnTheLeft = true;
-            }
-        }
-
-        return new Pair<>(hasNotchOnTheLeft, hasNotchOnTheRight);
     }
 
     @UiThread
@@ -683,5 +233,51 @@ class InAppMessageView extends WebViewClient {
             return deepLink;
         }
     }
+
+
+    // - Implementations for abstracts
+
+    @Override
+    protected JSONObject getCurrentMessageContent() {
+        JSONObject content = currentMessage.getContent();
+        if (region != null) {
+            try {
+                content.put("region", region);
+            } catch (JSONException e) {
+                Log.w(TAG, "Could not pass region to In-App renderer");
+            }
+        }
+
+        return content;
+    }
+
+    @Override
+    protected void onMessageClosedByClient() {
+        // this happens when IAR client closed message
+        presenter.messageClosed();
+    }
+
+    @Override
+    protected void onMessageCloseRequested() {
+        // this happens when we told IAR to close message
+        // TODO: the split keeps existing behaviour, but simpler would be to run this when message closed by client as well (or was there a reason?)
+        InAppMessageService.handleMessageClosed(currentActivity, currentMessage);
+    }
+
+    @Override
+    protected void onMessageOpened() {
+        InAppMessageService.handleMessageOpened(currentActivity, currentMessage);
+    }
+
+    @Override
+    protected void onExecuteActions(JSONObject data) {
+        if (null == data) {
+            return;
+        }
+
+        List<ExecutableAction> actions = this.parseButtonActionData(data);
+        currentActivity.runOnUiThread(() -> this.executeActions(currentActivity, actions));
+    }
+
 
 }
